feat(触摸设置): 新增触摸设置相关页面和功能
新增触摸设置主页面、IC状态页面、灵敏度调整页面及交互式灵敏度调整功能
重构页面导航管理器使用完整历史路径替代单页面记录
扩展页面跳转宏支持传递额外字符串参数
调整主菜单和主页面布局，移除测试页面
优化配置管理器的流式JSON处理能力

+++ src/protocol/hid/hid.cpp
@@ -111,7 +111,7 @@ void HID::force_send_keyboard_report() {
 }
 
 // 获取实际回报速率
-uint8_t HID::get_report_rate() const {
+uint32_t HID::get_report_rate() const {
     return cached_report_rate_;
 }
